+++ Panconvert.py
@@ -19,10 +19,14 @@ __author__ = 'apaeffgen'
 
 import codecs
 import platform
+import glob
+import os
 
 from PyQt5 import QtWidgets
 from PyQt5 import QtCore
 from source.dialog_preferences import *
+from source.dialog_batch import *
+from source.dialog_info import *
 from source.converter.markdown import *
 from source.converter.latex import *
 from source.converter.opml import *
@@ -30,39 +34,68 @@ from source.converter.html import *
 from source.converter.epub import *
 from source.converter.lyx import *
 from source.converter.manual_converter import *
+from source.converter.batch_converter import *
 from source.gui.panconvert_gui import Ui_notepad
+
+global openfile, filelist
 
 #TODO: Copy/Paste. In Ui + Function
 #TODO: Batch-Conversion. In Ui + Function
 # noinspection PyStatementEffect,PyAttributeOutsideInit,PyAttributeOutsideInit,PyAttributeOutsideInit
 class StartQT5(QtWidgets.QMainWindow):
+    global openfile, filelist
 
     """File Dialog Functions"""
 
     def file_dialog(self):
-        global text, data, openfile, path_dialog
+        global text, data, openfile, filelist, path_dialog
+
+        batch_settings = QSettings('Pandoc', 'PanConvert')
+        batch_open_path = batch_settings.value('batch_open_path')
+        batchConversion = self.ui.BatchConversion.isChecked()
+
+        if batchConversion is False:
+
+            fd = QtWidgets.QFileDialog(self)
+            if path_dialog == '':
+                fd.setDirectory(QtCore.QDir.homePath())
+            else:
+                fd.setDirectory(path_dialog)
 
 
-        fd = QtWidgets.QFileDialog(self)
-        if path_dialog == '':
-            fd.setDirectory(QtCore.QDir.homePath())
-        else:
-            fd.setDirectory(path_dialog)
+            self.filename = fd.getOpenFileName()
+            openfile = self.filename[0]
+            from os.path import isfile
+            if isfile(self.filename[0]):
+                try:
+                    text = codecs.open(self.filename[0], 'r', 'utf-8').read()
+                    data = self.ui.editor_window.setPlainText(text)
+
+                except:
+                    text = 'No Preview of the File-Data possible. Try to manually convert. Good Luck.'
+                    data = self.ui.editor_window.setPlainText(text)
+                    #QtWidgets.QMessageBox.warning(None, 'Warning-Message',
+                    #                          'No Preview of the File-Data possible. Try to manually convert. Good Luck.')
+
+        elif batchConversion is True:
+
+            fd = QtWidgets.QFileDialog(self)
+
+            if batch_open_path == '':
+                fd.setDirectory(QtCore.QDir.homePath())
+            else:
+                fd.setDirectory(batch_open_path)
 
 
-        self.filename = fd.getOpenFileName()
-        openfile = self.filename[0]
-        from os.path import isfile
-        if isfile(self.filename[0]):
-            try:
-                text = codecs.open(self.filename[0], 'r', 'utf-8').read()
-                data = self.ui.editor_window.setPlainText(text)
+            self.filename = fd.getOpenFileNames()
+            openfile = self.filename[0]
 
-            except:
-                text = 'No Preview of the File-Data possible. Try to manually convert. Good Luck.'
-                data = self.ui.editor_window.setPlainText(text)
-                #QtWidgets.QMessageBox.warning(None, 'Warning-Message',
-                #                          'No Preview of the File-Data possible. Try to manually convert. Good Luck.')
+            data = ('\n').join(openfile)
+
+            self.ui.editor_window.appendPlainText(data)
+            files = self.ui.editor_window.toPlainText()
+            filelist = files.split('\n')
+
 
     def file_save(self):
         from os.path import isfile
@@ -110,6 +143,21 @@ class StartQT5(QtWidgets.QMainWindow):
         text = text_undo
         self.ui.editor_window.setPlainText(text_undo)
 
+    def list_from_formats(self):
+        from_formats, to_formats = get_pandoc_formats()
+        QtWidgets.QMessageBox.information(None, 'Warning-Message',
+                                          'List of From-Formats: <br><br>'+ ', '.join(from_formats))
+
+    def list_to_formats(self):
+        from_formats, to_formats = get_pandoc_formats()
+        QtWidgets.QMessageBox.information(None, 'Warning-Message',
+                                          'List of To-Formats: <br><br>'+ ', '.join(to_formats))
+
+    def list_options(self):
+        options =  get_pandoc_options()
+
+        QtWidgets.QMessageBox.information(None, 'Warning-Message', 'Detailed-Text'
+                                          'List of Options: <br><br>'+ ', '.join(options))
 
     '''Export Functions'''
 
@@ -270,6 +318,19 @@ class StartQT5(QtWidgets.QMainWindow):
             QtWidgets.QMessageBox.warning(None, 'Warning-Message',
                                           'You have no Data to be converted. Please make an input')
 
+
+
+    def file_export_markdown2epub(self):
+        global text
+        text = self.ui.editor_window.toPlainText()
+        if text is not "":
+            output_content = convert_md2epub(text)
+            self.ui.editor_window.setPlainText(output_content)
+            text = output_content
+        else:
+            QtWidgets.QMessageBox.warning(None, 'Warning-Message',
+                                          'You have no Data to be converted. Please make an input')
+
     def file_export_manualconverter(self):
 
         global text, text_undo, openfile
@@ -302,21 +363,66 @@ class StartQT5(QtWidgets.QMainWindow):
                                           'Something went terribly wrong. Please get some help. Google never was your'
                                           '/ is your friend. Just the NSA is.')
 
-    def file_export_markdown2epub(self):
-        global text
-        text = self.ui.editor_window.toPlainText()
-        if text is not "":
-            output_content = convert_md2epub(text)
-            self.ui.editor_window.setPlainText(output_content)
-            text = output_content
+    def file_export_batch_conversion_standard(self):
+        # global openfile
+        # output_content = batch_convert(openfile,toFormat,fromFormat)
+        # self.ui.editor_window.setPlainText(output_content)
+        QtWidgets.QMessageBox.warning(None, 'Warning-Message',
+                                          'The batch conversion at the moment only works with the manual conversion.<br><br>'
+                                          ' Please uncheck the checkbox "Standard Conversion".')
+
+
+    def file_export_batch_conversion_manual(self):
+        global openfile, filelist
+
+        batch_settings = QSettings('Pandoc', 'PanConvert')
+        batch_convert_files = batch_settings.value('batch_convert_files')
+        batch_convert_directory = batch_settings.value('batch_convert_directory')
+        batch_convert_recursive = batch_settings.value('batch_convert_recursive')
+        data = self.ui.editor_window.toPlainText()
+
+
+
+        if data is not '' and batch_convert_files is True:
+
+            for openfiles in filelist:
+                if os.path.isfile(openfiles) is True:
+                    self.ui.editor_window.setPlainText(data)
+                    message, error = batch_convert_manual(openfiles,fromFormat,toFormat,extraParameter)
+
+                    self.ui.editor_window.appendPlainText(error)
+                    self.ui.editor_window.appendPlainText(message)
+                else:
+                    errormessage = ('Einige Dateiangaben waren nicht korrekt:')
+                    self.ui.editor_window.appendPlainText(errormessage)
+
+        elif batch_convert_recursive is False and batch_convert_directory is True:
+            batch_open_path = batch_settings.value('batch_open_path')
+
+            filelist = glob.glob(batch_open_path + '/*')
+            for openfiles in filelist:
+                if os.path.isfile(openfiles):
+                    output_content = batch_convert_manual(openfiles,fromFormat,toFormat,extraParameter)
+                    self.ui.editor_window.setPlainText(output_content)
+
+        elif batch_convert_recursive is True and batch_convert_directory is True:
+            batch_open_path = batch_settings.value('batch_open_path')
+
+            filelistrecursive = create_filelist(batch_open_path)
+
+            for openfiles in filelistrecursive:
+                output_content = batch_convert_manual(openfiles,fromFormat,toFormat,extraParameter)
+                self.ui.editor_window.setPlainText(output_content)
+
         else:
             QtWidgets.QMessageBox.warning(None, 'Warning-Message',
-                                          'You have no Data to be converted. Please make an input')
+                                              'You have to open at least one file in file conversion mode')
+
 
     """Gui-Event Definitions"""
 
     def __init__(self, parent=None):
-        global path_dialog, text, text_undo
+        global path_dialog, text, text_undo, openfile
 
         text = ''
         text_undo = ''
@@ -329,6 +435,9 @@ class StartQT5(QtWidgets.QMainWindow):
         self.ui.actionSave_AS.triggered.connect(self.file_save_as)
         self.ui.actionNew.triggered.connect(self.file_new)
         self.ui.actionUndo.triggered.connect(self.undo)
+        self.ui.ButtonFromFormat.clicked.connect(self.list_from_formats)
+        self.ui.ButtonToFormat.clicked.connect(self.list_to_formats)
+        self.ui.ButtonOptions.clicked.connect(self.info_dialog)
         self.ui.actionMarkdown2Latex.triggered.connect(self.file_export_md2latex)
         self.ui.actionOpml2latex.triggered.connect(self.file_export_opml2latex)
         self.ui.actionLatex2Opml.triggered.connect(self.file_export_latex2opml)
@@ -342,9 +451,10 @@ class StartQT5(QtWidgets.QMainWindow):
         self.ui.actionHtml2Latex.triggered.connect(self.file_export_html2latex)
         self.ui.actionLatex2html.triggered.connect(self.file_export_latex2html)
         self.ui.actionMarkdown2Lyx.triggered.connect(self.file_export_markdown2lyx)
+        self.ui.ButtonBatch.clicked.connect(self.batch_dialog)
         self.ui.ButtonRevert.clicked.connect(self.undo)
         self.ui.ButtonConvert.clicked.connect(self.event_triggered)
-        self.ui.actionPreferences.triggered.connect(self.event_dialog)
+        self.ui.actionPreferences.triggered.connect(self.preference_dialog)
         self.ui.actionHelp.triggered.connect(self.help_dialog)
         self.ui.actionPandoc_Docs.triggered.connect(self.pandocdocs_dialog)
         self.ui.actionSave.setEnabled(True)
@@ -358,6 +468,7 @@ class StartQT5(QtWidgets.QMainWindow):
         xtraParameter = settings.value('xtraParameter')
 
         Standard_Conversion = settings.value('Standard_Conversion', False)
+        Batch_Conversion = settings.value('Batch_Conversion', False)
         From_Markdown = settings.value('From_Markdown', False)
         From_Html = settings.value('From_Html', False)
         From_Latex = settings.value('From_Latex', False)
@@ -373,6 +484,7 @@ class StartQT5(QtWidgets.QMainWindow):
         if settings.value('From_Markdown') is not None:
             if platform.system() == 'Windows' or platform.system() == 'Linux':
                 self.ui.StandardConversion.setChecked(strtobool(Standard_Conversion))
+                self.ui.BatchConversion.setChecked(strtobool(Batch_Conversion))
                 self.ui.ButtonFromMarkdown.setChecked(strtobool(From_Markdown))
                 self.ui.ButtonFromHtml.setChecked(strtobool(From_Html))
                 self.ui.ButtonFromLatex.setChecked(strtobool(From_Latex))
@@ -388,6 +500,7 @@ class StartQT5(QtWidgets.QMainWindow):
 
             else:
                 self.ui.StandardConversion.setChecked(Standard_Conversion)
+                self.ui.BatchConversion.setChecked(Batch_Conversion)
                 self.ui.ButtonFromMarkdown.setChecked(From_Markdown)
                 self.ui.ButtonFromHtml.setChecked(From_Html)
                 self.ui.ButtonFromLatex.setChecked(From_Latex)
@@ -413,10 +526,20 @@ class StartQT5(QtWidgets.QMainWindow):
         import webbrowser
         webbrowser.open_new_tab('http://pandoc.org/README.html')
 
-    def event_dialog(self):
+    def preference_dialog(self):
         """ References to dialog_preferences.py"""
         self.PreferenceDialog = PreferenceDialog(self)
         self.PreferenceDialog.show()
+
+    def batch_dialog(self):
+        """ References to dialog_batch.py"""
+        self.BatchDialog = BatchDialog(self)
+        self.BatchDialog.show()
+
+    def info_dialog(self):
+        """ References to dialog_batch.py"""
+        self.InfoDialog = InfoDialog(self)
+        self.InfoDialog.show()
 
 
 
@@ -428,8 +551,9 @@ class StartQT5(QtWidgets.QMainWindow):
         toFormat = self.ui.ToParameter.text()
         standard_conversion = self.ui.StandardConversion.isChecked()
         extraParameter = self.ui.ExtraParameter.text()
+        batchConversion = self.ui.BatchConversion.isChecked()
 
-        if standard_conversion is True:
+        if standard_conversion is True and batchConversion is False:
             if self.ui.ButtonFromMarkdown.isChecked() is True and self.ui.ButtonToLatex.isChecked() is True:
                 self.file_export_md2latex()
             elif self.ui.ButtonFromMarkdown.isChecked() is True and self.ui.ButtonToOpml.isChecked() is True:
@@ -461,14 +585,24 @@ class StartQT5(QtWidgets.QMainWindow):
                                               'The from-Format and to-Format should not be identical.<br><br> '
                                               'If you picket to-Lyx, only from-markdown is a valid option.<br><br>'
                                               'Please make a different choice.')
+
+        elif standard_conversion is False and batchConversion is False:# and extraParameter is not "":
+            self.file_export_manualconverter()
+
+        elif standard_conversion is True and batchConversion is True:
+            self.file_export_batch_conversion_standard()
+
+
+        elif standard_conversion is False and batchConversion is True:
+            self.file_export_batch_conversion_manual()
+
         elif fromFormat is "" or toFormat is "":
             QtWidgets.QMessageBox.warning(None, 'Warning-Message',
                                           'If you fill in Arguments and uncheck the Box "Standard", you have to '
                                           'provide at least the following Parameters: From, To. <br><br>'
                                           '  Some Formats like odt, epub need an input '
                                           'for "Parameter". Otherwise there will be no output at all')
-        elif fromFormat is not "" and toFormat is not "":# and extraParameter is not "":
-            self.file_export_manualconverter()
+
 
         else:
             QtWidgets.QMessageBox.warning(None, 'Warning-Message',